fix: prevent closing modal during language transfer
